Added statistical analysis and a full Pokedex scraper
The Pokedex scraper rips data from Pokemondb.com and store it in a .json file.  The format is the same as before, but there are all Pokemon and forms in the file now.  The stats_analysis.py file currently only finds the average and standard deviation of all stats, but more advanced features will be added in the future.